# perf: Add AES key caching