… go.sum
… testdata/aes7z.7z