Before:
```
$ go test -v -run '^$' -bench='AES'
goos: darwin
goarch: amd64
pkg: github.com/bodgit/sevenzip
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkAES7z
BenchmarkAES7z-12    	       2	 533112205 ns/op
PASS
ok  	github.com/bodgit/sevenzip	1.808s
```

After:
```
$ go test -v -run '^$' -bench='AES'
goos: darwin
goarch: amd64
pkg: github.com/bodgit/sevenzip
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkAES7z
BenchmarkAES7z-12    	    1167	   1063848 ns/op
PASS
ok  	github.com/bodgit/sevenzip	2.427s
```

Fixes #187